ArVirtual#select_from_alias - use proper alias
before:
We are comparing the to and the to model.
No surprise, it always was the same and always generated a new alias.

after:
Compare the to and the src.
Only pass 1 reference for the to table
Fix documentation error

(transferred from ManageIQ/manageiq@1e6304280838d481646f82f09b4fb1b75e27d13c)