Fixes virtual_totals when a default_scope exists
When a has_many relation of a model that is used in a virtual_total and
has a default scope, it was not getting applied to the arel query when
the virtual_total column was being used in the select.  An example in
CloudTenant:

    class CloudTenant < ApplicationRecord

      has_many   :vms, -> { active }

      virtual_total :total_vms, :vms

      # ...
    end

Then flowing ActiveRecord query:

    CloudTenant.where(:name => 'DEV').select(:total_vms)

Would produce:

    SELECT (
      (SELECT COUNT(*)
       FROM "vms"
       WHERE "cloud_tenants"."id" = "vms"."cloud_tenant_id")
      ) AS total_vms
    FROM "cloud_tenants"
    WHERE "cloud_tenants"."name" = 'DEV'

Which doesn't take into account the `active` scope, which will filter
out any VM without an `ems_id` (so templates.

With this change, the SQL executed then gets updated to the following:

    SELECT (
      (SELECT COUNT(*)
       FROM "vms"
       WHERE "cloud_tenants"."id" = "vms"."cloud_tenant_id"
         AND "vms"."type" IN ('Vm', 'ManageIQ::Providers::CloudManager::Vm',
                              'ManageIQ::Providers::InfraManager::Vm', 'VmServer',
                              'ManageIQ::Providers::Amazon::CloudManager::Vm',
                              'ManageIQ::Providers::Azure::CloudManager::Vm',
                              'ManageIQ::Providers::Google::CloudManager::Vm',
                              'ManageIQ::Providers::Openstack::CloudManager::Vm',
                              'ManageIQ::Providers::Vmware::CloudManager::Vm', 'VmXen',
                              'ManageIQ::Providers::Redhat::InfraManager::Vm',
                              'ManageIQ::Providers::Microsoft::InfraManager::Vm',
                              'ManageIQ::Providers::Vmware::InfraManager::Vm')
         AND ("vms"."ems_id" IS NOT NULL))
       ) AS total_vms
    FROM "cloud_tenants"
    WHERE "cloud_tenants"."name" = 'DEV'

Which properly includes the default scopes for both the `Vm` model, and
the `:vm` relation in `CloudTenant`.


(transferred from ManageIQ/manageiq@55760b37d1a310575c957cb624d907aa2cd35ea4)