Fix virtual_aggregate_arel with default_scope
`.scope_for` on a `ActiveRecord::Reflection` doesn't actually return the
default scope, but calls `unscoped` on the `klass` of the reflection,
and returns that.  This is not wanted in this case since we expect the
`default_scope` to be there.

The other issue with this is sending a raw `.ast` from a where clause
will pass an `arel` object, but it will still have the bind parameters
removed when it is executed so:

    'vms'.'template' = $1

instead of what would be expected:

     'vms'.'template' = 'f'

This is because Rails/Arel only bind the attributes after they have been
converted to a SQL string, instead of when they are in the `node` form.

To get around this, we are instead building an ActiveRecord query based
on the relations' default scope (and the scope of the relation, if it is
present), instead of building a new query, calling `.to_sql` on it,
which will bind the attributes to the SQL string, and then passing that
as a `Arel::Nodes::SqlLiteral` to the `t.grouping` instead.

This assures that the `default_scope` on the relation, the scope of the
relation itself, and the binded params of those two scopes are all in
place when building the `virtual_total` arel.  This is a bit hacky, and
makes the resulting arel from that column static and unable to be
modified, but we shouldn't need to work with the `virtual_total` arel
directly when passing it as columns for `.select` and `.order`.

This also handles cases where `.order` is used in the relation, since it
will most likely fail since it is not used in the aggregation function
all of the time.  Since this is an aggregation function, though, this
shouldn't be a problem, or used.

So this all should be okay... I think...

Added a bunch of tests to confirm this functionality works.


(transferred from ManageIQ/manageiq@f8dddabc46991f09ff6988a4cc812b2cdf832573)