use aggregates from attributes
before: if the aggregate is nil, it used to perform an extra query

problem: attribute_present? was false for a nil valued attribute
fix: has_attribute is true for a nil valued attribute
problem: fell back to lookup clause for nil value
fix: falls back to default clause instead (0)

As for size, select count(*) will always return a number,
  so the attribute can't be nil
and in ruby [].size will always return a number.
  so || 0 is probably not necessary nor is the try
But left them there incase it is a delegate or some special case due to
  a bug in development