virtual_aggregate nil attribute does not exec query
attribute_present? is basically has_attribute? and !value.nil? combined

Changing to has_attribute? to detect that the attribute is present.
In the nil case, then it doesn't make sense to run the query again since it would
just be nil again. Instead, it just returns the default.

But you will notice that the return value changes.
That is because the default value is not always 0. It is sometimes 0
and other times nil.

So while there is change, the results are more consistent and not dependent
upon whether a relation is preloaded or not.