for select, map fields at query time
currently we monkey patch select() because there are some times
when the code will not properly fetch the arel for an attribute.
Specifically when we specify a database name, though there may
be some other edge cases.

Instead of changing the column in the select up front, we are now
passing it through as typical rails works and fixing the core
to properly recognized virtual attributes.

Most of the changes going into rails 5.2 and 6.0 involve less
monkey patching. This, in contrast, patches deeper in the code,
and requires a bit of copying code from rails core.

While this is not desirable, it is best to keep the columns as
symbols until the end.