add virtual_aggregate :size back in
a virtual_total is basically a virtual_aggregate for the size
function.

Rails handles count(*) differently than sum(column).
The two methods were split apart to separate the code since
when they were in one method, there were a bunch of if branches

But it is possible that someone would get a little confused, so
adding support to allow people to call virtual_aggregate and
pass in size as the function.