No longer hard code size method in arel generation
Since creating aggregate attribute arel has 2 paths, there is no need
to pass in a method of size and then change it to count.
Instead, passing in a method of count and remove the check.

The field itself is simpler. If there is an arel friendly attribute,
leave it as arel (checking `respond_to?(:count)`), otherwise convert
to an arel attribute