work around attribute.name
Rails is now calling:
attribute = arel_attribute()
attribute.name

All of our arel_attributes return a grouping or an attribute

This change allows our return values from arel_attributes
to respond to grouping.name, so serialization information can be looked
up in rails.

I looked into deferring the call to arel_attributes in rails, but this
requires a change in the attribute handlers, which has been an exposed
interface for a while, so it is not possible.
I'm leaving as monkey patch stage.

We may want to look into getting closer to the attribute code

Not thrilled with this solution.
Can't subclass Grouping because class is hardcoded.
Can't subclass Attribute for same reason

introduced by https://github.com/rails/rails/pull/32981