no longer support invalid virtual_column calls
The `virtual_column` method was sometimes called with a type instead of `:type =>`.
While this was never valid, it was prevalent enough to code a fix.

These have all been cleared up, and this PR officially drops support for the bad call.

we're dropping support because ruby 2.7 doesn't like
a hash and named keys right next to each other.

This removes the warning:

```
warning: Passing the keyword argument as the last hash parameter is deprecated
```