extend through arel_table instead of arel_attribute
arel_table[] only accepted a single attribute before 6.1
So when running 6.0, this also mimics the call to arel_attribute

NOTE: arel_table[] now resolved aliases for delegation

This is a deviation from rail 6.1
In 6.0, we always went through arel_attribute (that did have this behavior)
A few parts are acting up because they want the alias resolution code.

Moving this code to arel_table helps in a few edge cases. (there are tests for these)