Monkeypatch Arel::HomogeneousIn
6.1 introduced HomogeneousIn, an optimization for In clauses.
Typically nodes like this will visit the node on the left to produce
the sql for the attribute, but instead this generates the raw sql for
the attribute as well.

In our case, it is expecting a simple node, so it does not generate the sql
correctly.

I put in a PR for rails to visit the left node, but in the mean time we
require this monkey patch.