add type for virtual_delegate tests
We used to just derive the types for virtual_delegates from the target attribute.
This rarely caused deadlocks from a race conditions in the class loader.
We removed all uses of this feature.

I am removing this from the test cases.

Still left the code that derives the type, but it is discouraged.