spec: ensure delegates work with OR and HomogeneousIn
mostly focused on ensuring the arel coming from the delegate works correctly.

I thought we had the grouping and to_sql in VirtualDelegate#select_from_alias
was to ensure that binding variables did not get botched,
but I can't reproduce the issue and get it to fail.

Adding these specs for now and hoping to get a good reproducer for the need
of the Arel.sql(arel.to_sql)