Use database_cleaner to clear out tables
We would like to clean up all records created in each test.
This was not properly working before so we had to manually delete records ourselves.

Rails ActiveRecord does provide  transactional_features for tests, but it is not
so simple to implement. This functionality is provided by `rspec-rails`, but not
in standard `rspec`.

Using DatabaseCleaner gem to clean up after each test.
The `transaction` strategy should work for us since we only use a single database
and a single connection.