support bind variables in virtual delegate reflections
We were not properly escaping the arel in the association de-reference.

This use case is very particular and will only happen with hard coded
strings in associations.

13016ba214dbfd98 rewrote to_sql to wrap in unprepared_statement in one
spot but not in this spot. So I guess this continued that effort.

a53d9597d4550345 stated there was a problem but had no reproducer
This added specs, but they don't look like they would be able to detect
the mentioned issue. So I guess this continued that effort as well.