Optimization for an includes that is a single key
no reason to convert the key into a hash and then run a deep merge on it.
Instead, just adding it to the hash.

NOTE: the string case is still flowing through the normal flow. But since
Strings are not common, it didn't seem necessary to add more code here