replace_virtual_field condenses returning includes
Merged_includes via include_to_hash already promotes classes up to a Hash.
But there was no code that demoted classes back to an array or symbol / nil.

If you look at the changed specs, you can see how this reduces the values.

In the rails 7.0 migration, I was going to add special code to introduce this
but it seems best to do this at the source