Introduce ActiveRecord::VirtualAttributes::VirtualAttribute
We used to override Grouping node.
It represents a parenthesized grouping of sql, which is fitting.

But since we also need to handle an attribute name and type casting, it is
time to graduate to having our own node.

Still leveraging the Grouping visitor in ToSql.

There are a few is_a?(Grouping), so added tests around the nodes where I
felt that may come up.