Don't run preloader on double arrays
This is an optimization.

`klass_associations` needs to be an array.
It can be a `nil, symbol, Array` or `Hash`.

The `nil` case is handled by `blank?`.
The `Symbol` and `Hash` are converted to an `Array`
But the `Array` was being converted to a double `Array`.

Which was not horrible, but just calls a loader then another loader.

With this change, the loader only runs through the association once.