Pull define_virtual_{include,arel} out of define_virtual_attribute
instead of defining arel and includes at runtime, we're defining them up front.

Introducing the arel and includes up front simplifies the virtual attribute definition
to just registering with the attributes api.

If you note, the code only uses half of the attribute call, specifically `attribute_type`.
This is causing problems in 7.2.

We simplified it to here so it will be easier in the transition to 7.2