Drop virtual_delegates_to_define
By defining the ruby method up front and registering the arel and uses,
the code can use the new simplified virtual attribute registration process.

Care does need to go into ensuring that virtual_delegate_arel does not access the schema

Do note, we are not passing options into virtual_attributes, which get passed to Type.lookup.
This was not working for virtual_delegates before.
TBH/ I don't know if we use this in the type lookup. Maybe these are be used when using array types?
It looks like rails uses the options for the matchers. It wasn't fully clear for me.