Introduce virtual_has_many through
This will automatically add the ruby deletate

It is assumed that a nil association is allowed, and a default of [] will be returned.

The delegate defined does still have logic for attributes (namely the has_attribute?) part.
Since there will be no attribute with this name, it will be ignored.

Not using traditional `delegate`, because that does not provide the `default` option