Support rails 7.2
https://github.com/rails/rails/commit/e0a55b038f7f2f50d1467876558be183be6cedaa

While we do not leverage the `attribute` api, we are strongly influenced by it.
We do not want a default value record, which modifies select(*) calls and others.
We want to add the values to `attribute_type`.

Before
------

The values are stored in `attributes_to_define_after_schema_loads` at class load time,
and in load_schema, are then used to generate entries for `DefaultValue` and `attribute_type`.

After
-----

The values are stored as `ModifyAttribute` entries.
Those values then are used to populate default values and attribute_type on demand.

Our Change
----------

Previous PRs removed arel and uses logic so load_schema only manipulated attribute_type.
Like the rails change, we populate (modify) attribute_type on demand.