drop unnecessary load_schema statements
Before rails 7.2:
All of this data was set in `load_schema!` so we needed to ensure
the schema was loaded to access _virtual_includes or _virtual_arel.
Same may be the case for has_attribute?

After rails 7.2:
It may have always been the case, but at least for the rails methods
we do not need to ensure the schema is loaded.
For our 2 class level variables, we now load _virtual_arel without
load_schema. So there is no reason to wait for schema_load.

- virtual_attribute?:
  - has attribute loads the schema if necessary
  - column_for_attribute loads the schema if necessary
- attribute_supported_by_sql?
  - has_attribute loads the schema if necessary
  - virtual_attribute? loads the schema indirectly (by 2 internal methods)
  - _virtual_arel is populated on class load, not schema load
- virtual_includes
  - _virtual_includes is populated on class load, not schema load